requirements.txt
PySide6>=6.5.0
Pillow>=10.0.0  # PIL library for image processing
python-dateutil>=2.8.2 